+++ CHANGELOG.md
@@ -23,7 +23,7 @@ All notable changes to this project will be documented in this file. This projec
 - Prevented app window from expanding too much horizontally when processing extremely long filename or URL inputs.
 - Fixed incorrect `[0m` that terminated script output in app.
 - Fixed MacUpdate description pattern matching.
-- Better handling of apps (and their associated URLs) with spaces, ampersands, and other special characters.
+- Better handling of URLs with unquoted spaces. (#197)
 - Improved detection of less obvious Sparkle feeds that might at first glance appear to be download URLs.
 
 ### Changed